jni: added Android Studio integration for quick NDK development
Since feza is a gcc environment for normal linux use, it is
adjustable for Android Studio C/C++ Native Development Kit. The
idea is to create a jni shim layer that communicates between the
Android Studio and the current feza project.

This patch adds the jni shim layer after tested working with
Android Studio 3.0.1 default framework.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>